test: add invoice test to storj-up
Adding a basic invoicing test. This test will create test data to generate an invoice, and apply a partial payment to the invoice amount via storjscan native token payment. Valid stripe public and secret keys are required.

+++ test/test-invoice.sh
@@ -16,6 +16,9 @@ fi
 export STORJUP_NO_HISTORY=true
 
 storj-up init storj,db,billing
+storj-up env setenv satellite-api satellite-core satellite-admin STORJ_PAYMENTS_PROVIDER=stripecoinpayments
+storj-up env setenv satellite-api satellite-core satellite-admin STORJ_PAYMENTS_STRIPE_COIN_PAYMENTS_STRIPE_PUBLIC_KEY="$STRIPE_PUBLIC_KEY"
+storj-up env setenv satellite-api satellite-core satellite-admin STORJ_PAYMENTS_STRIPE_COIN_PAYMENTS_STRIPE_SECRET_KEY="$STRIPE_SECRET_KEY"
 
 docker compose down -v
 docker compose up -d
@@ -42,12 +45,37 @@ curl -X POST 'http://localhost:10000/api/v0/payments/wallet' --header "$COOKIE"
 ADDRESS=$(curl -X GET -s http://localhost:10000/api/v0/payments/wallet --header "$COOKIE" | jq -r '.address')
 
 #ACCOUNT is defined with environment variables above
-for i in {1..15}; do cethacea token transfer 1000 0x"$ADDRESS"; done
+for i in {1..15}; do cethacea token transfer 10 0x"$ADDRESS"; done
 
 # todo: find a better way than sleep to wait until token balance chores reflect in billing table
 sleep 180
 
 curl -X GET http://localhost:10000/api/v0/payments/wallet --header "$COOKIE"
+curl -X POST http://localhost:10000/api/v0/payments/account --header "$COOKIE"
+
+# invoicing
+storj-up testdata project-usage
+storj-up testdata fix-billing
+YEAR=$(date +%Y)
+if [ $(uname -s) == "Darwin" ]
+then
+  MONTH=$(date -v-1m +%m)
+else
+  MONTH=$(date +%m -d 'last month')
+fi
+
+docker-compose exec satellite-admin satellite billing prepare-invoice-records "$MONTH"/"$YEAR" --log.level=info --log.output=stdout
+docker-compose exec satellite-admin satellite billing create-project-invoice-items "$MONTH"/"$YEAR" --log.level=info --log.output=stdout
+docker-compose exec satellite-admin satellite billing create-invoices "$MONTH"/"$YEAR" --log.level=info --log.output=stdout
+docker-compose exec satellite-admin satellite billing create-token-invoice-items "$MONTH"/"$YEAR" --log.level=info --log.output=stdout
+docker-compose exec satellite-admin satellite billing finalize-invoices --log.level=info --log.output=stdout
+
+BALANCE=$(curl -X GET -s http://localhost:10000/api/v0/payments/wallet --header "$COOKIE" | jq -r '.balance')
+
+if [[ $BALANCE == -* ]]
+then
+  exit 1
+fi
 
 docker compose down
 rm -rf .contracts.yaml
